+++ small_text/query_strategies/bayesian.py
@@ -4,7 +4,7 @@ from typing import Union
 from scipy.sparse import csr_matrix
 from small_text.classifiers import Classifier
 from small_text.data import Dataset
-from small_text.query_strategies.base import QueryStrategy
+from small_text.query_strategies.base import QueryStrategy, ScoringMixin
 
 
 def _bald(p, eps=1e-8):
@@ -14,7 +14,7 @@ def _bald(p, eps=1e-8):
     return model_prediction_entropy - expected_prediction_entropy
 
 
-class BALD(QueryStrategy):
+class BALD(ScoringMixin, QueryStrategy):
     """Selects instances according to the Bayesian Active Learning by Disagreement (BALD) [HHG+11]_
     strategy.
 
@@ -32,6 +32,25 @@ class BALD(QueryStrategy):
             to obtain uncertainty estimates.
         """
         self.dropout_samples = dropout_samples
+        self.scores_: Union[npt.NDArray[np.double], None] = None
+
+    @property
+    def last_scores(self):
+        return self.scores_
+
+    def score(self,
+              clf: Classifier,
+              dataset: Dataset,
+              indices_unlabeled: npt.NDArray[np.uint],
+              indices_labeled: npt.NDArray[np.uint],
+              y: Union[npt.NDArray[np.uint], csr_matrix]) -> npt.NDArray[np.double]:
+        _unused = indices_unlabeled, indices_labeled
+        proba_dropout_sampled = clf.predict_proba(dataset, dropout_sampling=self.dropout_samples)
+
+        bald_scores = _bald(proba_dropout_sampled)
+        self.scores_ = bald_scores
+
+        return bald_scores
 
     def query(self,
               clf: Classifier,
@@ -42,8 +61,7 @@ class BALD(QueryStrategy):
               n: int = 10) -> np.ndarray:
         self._validate_query_input(indices_unlabeled, n)
 
-        proba_dopout_sampled = clf.predict_proba(dataset, dropout_sampling=self.dropout_samples)
-        bald_scores = _bald(proba_dopout_sampled)
+        bald_scores = self.score(clf, dataset, indices_unlabeled, indices_labeled, y)
 
         if len(indices_unlabeled) == n:
             return np.array(indices_unlabeled)