+++ tests/utils/testing_smalltext.py
@@ -1,0 +1,7 @@
+from small_text.data.datasets import DatasetView
+
+
+class AnyDatasetView(object):
+
+    def __eq__(self, other):
+        return isinstance(other, DatasetView)